Move time format function to utils #121
