+++ frontend/src/utils/timeFormat.js
